+++ src/pages/SignInPage/index.ts
@@ -1,0 +1,3 @@
+import SignInPage from "./SignInPage";
+
+export default SignInPage;
