+++ src/pages/ToDoPage/ToDoPage.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useReducer, useRef, useState} from 'react';
 import {RouteComponentProps} from 'react-router-dom';
 
-import reducer, {initialState} from './store/reducer';
+import reducer, {initialState} from 'reducers/todoList';
 import {
     setTodos,
     createTodo,
@@ -9,10 +9,10 @@ import {
     toggleAllTodos,
     deleteAllTodos,
     updateTodoStatus
-} from './store/actions';
-import Service from './service';
-import {TodoStatus} from './models/todo';
-import {isTodoCompleted} from './utils';
+} from 'actions/TodoListAction';
+import Service from 'service';
+import {TodoStatus} from 'models/todo';
+import {isTodoCompleted} from 'utils';
 
 type EnhanceTodoStatus = TodoStatus | 'ALL';
