+++ src/pages/ToDoPage/index.ts
@@ -1,0 +1,3 @@
+import ToDoPage from "./ToDoPage";
+
+export default ToDoPage;
